Modularize services and fix passive metadata typing
