+++ packages/engine/src/services/services_types.ts
@@ -1,0 +1,13 @@
+import type { ResourceKey } from '../state';
+import type { HappinessTierDefinition } from './tiered_resource_types';
+
+export type RuleSet = {
+	defaultActionAPCost: number;
+	absorptionCapPct: number;
+	absorptionRounding: 'down' | 'up' | 'nearest';
+	tieredResourceKey: ResourceKey;
+	tierDefinitions: HappinessTierDefinition[];
+	slotsPerNewLand: number;
+	maxSlotsPerLand: number;
+	basePopulationCap: number;
+};
